bunch of code deduplication and cleanup
